jupyter/requirements.txt
google-cloud-bigquery
google-cloud-bigquery-storage
pandas-gbq
pyarrow
db-dtypes
